fix(composes): 能力过滤改为「未声明即未知」+ sticky 取模越界 + 错误文档同步
评审反馈修复:

1. 【阻断】能力过滤默认「拒绝」破坏既有 tools/vision 调用。
   仓库内没有任何生产类型实现 CapabilityProvider(*aimodel.Client 也没有),
   未声明 Capability 的 entry 回落到零值后会被 tools 请求全部过滤掉,
   一次网络请求都不发。改为:resolvedCapability 返回 (Capability, declared),
   只有显式声明(非 nil ModelEntry.Capability 或 CapabilityProvider 客户端)
   才参与过滤;未声明 = 未知,永不排除。「无 tools 端点 + 有 tools 端点」
   场景仍成立(显式声明 Tools:false 即被排除且不降级)。

2. sticky 哈希在 32 位平台取到负数下标:int(h.Sum32()) 在 int 为 32 位时
   可能为负导致 aliases[idx] panic。改为在 uint32 空间内取模,并修正
   与代码不符的注释措辞。

3. 四路同步第 ③ 步:doc/design/errors.md §4/§5 改写,说明 compose 聚合错误
   现由 composes.MultiError/EndpointError 承担(含 errors.As 迁移写法),
   并明确 ais.ModelError/ais.MultiError 的去留(保留导出、仓库内无生产者)。

一并处理评审的三条可选改进:
- Stats() 按路由行为归一化状态:冷却期满的端点报 active,不再停留在 cooling;
- 429 应答恢复探测时不再把长退避降级为 10s 冷却,保持 error 态与当前退避档位;
- NewComposeClient 先克隆 entries 再派生 alias,不回写调用方切片。

新增单测锁住每条路径:未声明端点不被过滤(含端到端发送 tools)、声明与未声明
混排、sticky 下标恒在集合内、冷却归一化、429 不绕过退避、构造不改调用方切片。
doc/design/compose.md §1/§3/§4 同步更新。

+++ composes/sticky.go
@@ -68,9 +68,10 @@ func (c *ComposeClient) stickyPreferredAlias(sessionID string) string {
 		h.Write([]byte{0})
 	}
 
-	// Mix in the alias count via the digest domain to avoid modulo bias concerns
-	// being confused with ordering; the digest is stable for a given input.
-	idx := int(h.Sum32()) % len(aliases)
+	// Reduce the digest modulo the alias count in uint32 space: converting the
+	// uint32 digest to int first would yield a negative value (and a negative
+	// index) wherever int is 32 bits.
+	idx := int(h.Sum32() % uint32(len(aliases)))
 
 	return aliases[idx]
 }
